Fix line breaks within multi-line text
This hack allows paragraphs of multi-line text with manual line breaks to not
have line breaks in the copied text. Makes it very annoying to have random
newlines in the middle of text that looks like a paragraph.

This is a test implementation for two pages: index.html and criminal records.

Conflicts:
	criminal-records-lok-sabha/hi/index.html
Conflict due to fixes in spelling.